fix: move example to next-openai
