+++ .changeset/sour-mails-cheer.md
@@ -2,4 +2,4 @@
 'ai': patch
 ---
 
-chore (ai/mcp): add support for `sessionId` to experimental MCP client
+chore (ai/mcp): add support for `sessionId` and `strictMode` configuration to experimental MCP client
