+++ examples/mcp/package.json
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "@ai-sdk/openai": "1.3.19",
-    "@modelcontextprotocol/sdk": "^1.7.0",
+    "@modelcontextprotocol/sdk": "1.10.2",
     "ai": "4.3.10",
     "dotenv": "16.4.5",
     "express": "5.0.1",
