+++ examples/mcp/src/sse/server.ts
@@ -14,7 +14,6 @@ mcpServer.tool(
   'check-product-stock',
   'Check if a product is available',
   {
-    // @ts-ignore
     productName: z.string(),
   },
   async ({ productName }) => {
