+++ examples/mcp/src/stdio/server.ts
@@ -13,7 +13,6 @@ server.tool(
   'get-pokemon',
   'Get Pokemon details by name',
   {
-    // @ts-ignore
     name: z.string(),
   },
   async ({ name }) => {
