--- examples/next-mcp/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "next-mcp",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@ai-sdk/openai": "1.3.19",
-    "@ai-sdk/react": "1.2.9",
-    "@modelcontextprotocol/sdk": "1.10.2",
-    "ai": "4.3.10",
-    "next": "latest",
-    "react": "^18",
-    "react-dom": "^18",
-    "zod": "3.23.8"
-  },
-  "devDependencies": {
-    "@types/node": "20.17.24",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.4.14",
-    "eslint": "8.57.1",
-    "eslint-config-next": "14.2.3",
-    "postcss": "^8.4.49",
-    "tailwindcss": "^3.4.15",
-    "typescript": "5.6.3"
-  }
-}
