+++ examples/next-openai/app/mcp/server/route.ts
@@ -1,7 +1,8 @@
-import { mcpApiHandler } from '@/lib/handler';
-import { createServerResponseAdapter } from '@/lib/server-response';
+import { mcpApiHandler } from '@/util/mcp/handler';
+import { createServerResponseAdapter } from '@/util/mcp/server-response';
 import { NextRequest } from 'next/server';
 
+// This route (/mcp/server) serves the MCP server; it's called by the /mcp/chat route that's used by useChat to connect to the server and fetch tools:
 const requestHandler = (req: NextRequest) => {
   return createServerResponseAdapter(req.signal, res => {
     mcpApiHandler(req, res);
