+++ examples/next-openai/package.json
@@ -18,6 +18,7 @@
     "@ai-sdk/perplexity": "1.1.7",
     "@ai-sdk/ui-utils": "1.2.8",
     "@ai-sdk/react": "1.2.9",
+    "@modelcontextprotocol/sdk": "1.10.2",
     "@vercel/blob": "^0.26.0",
     "ai": "4.3.10",
     "next": "latest",
