+++ examples/next-openai/util/mcp/incoming-message.ts
