+++ examples/next-openai/util/mcp/server-response.ts
